modified:   minicas-beta.tar.gz 	modified:   src/include/num_q.h 	modified:   src/include/num_z.h 	modified:   src/include/num_zm.h 	modified:   src/include/number.h 	modified:   src/include/signatures.h 	modified:   src/num_q/attrib.cpp 	modified:   src/num_q/bool.cpp 	modified:   src/num_q/num_q.cpp 	modified:   src/num_q/op_atrdiv.cpp 	modified:   src/num_q/op_atrmul.cpp 	modified:   src/num_q/op_atrsub.cpp 	modified:   src/num_q/op_atrsum.cpp 	modified:   src/num_q/opdiv.cpp 	modified:   src/num_q/opmul.cpp 	modified:   src/num_q/opsub.cpp 	modified:   src/num_q/opsum.cpp 	modified:   src/num_z/attrib.cpp 	modified:   src/num_z/base.cpp 	modified:   src/num_z/bool.cpp 	modified:   src/num_z/io.cpp 	modified:   src/num_z/num_ops.cpp 	modified:   src/num_z/num_z.cpp 	modified:   src/num_z/op_atrdiv.cpp 	modified:   src/num_z/op_atrmod.cpp 	modified:   src/num_z/op_atrmul.cpp 	modified:   src/num_z/op_atrsub.cpp 	modified:   src/num_z/op_atrsum.cpp 	modified:   src/num_z/opdiv.cpp 	modified:   src/num_z/opmod.cpp 	modified:   src/num_z/opmul.cpp 	modified:   src/num_z/opsub.cpp 	modified:   src/num_z/opsum.cpp 	modified:   src/num_z/private_ops.cpp

+++ src/include/num_q.h
@@ -29,9 +29,9 @@
 			num_q(const num_q &);
 			num_q(const num_z &);
 			num_q(const num_z &, const num_z &);
-			num_q(const num_z &, const int64_t &);
-			num_q(const int64_t &, const num_z &);
-			num_q(const int64_t &, const int64_t &);
+			num_q(const num_z &, const long long &);
+			num_q(const long long &, const num_z &);
+			num_q(const long long &, const long long &);
 			num_q(int);
 			num_q(int, int);
 			
@@ -56,31 +56,31 @@
 			
 			num_q & operator=(const num_q &);
 			num_q & operator=(const num_z &);
-			num_q & operator=(const int64_t &);
+			num_q & operator=(const long long &);
 			num_q & operator+=(const num_q &);
 			num_q & operator+=(const num_z &);
-			num_q & operator+=(const int64_t &);
+			num_q & operator+=(const long long &);
 			num_q & operator-=(const num_q &);
 			num_q & operator-=(const num_z &);
-			num_q & operator-=(const int64_t &);
+			num_q & operator-=(const long long &);
 			num_q & operator*=(const num_q &);
 			num_q & operator*=(const num_z &);
-			num_q & operator*=(const int64_t &);
+			num_q & operator*=(const long long &);
 			num_q & operator/=(const num_q &);
 			num_q & operator/=(const num_z &);
-			num_q & operator/=(const int64_t &);
+			num_q & operator/=(const long long &);
 			num_q operator+(const num_q &) const;
 			num_q operator+(const num_z &) const;
-			num_q operator+(const int64_t &) const;
+			num_q operator+(const long long &) const;
 			num_q operator-(const num_q &) const;
 			num_q operator-(const num_z &) const;
-			num_q operator-(const int64_t &) const;
+			num_q operator-(const long long &) const;
 			num_q operator*(const num_q &) const;
 			num_q operator*(const num_z &) const;
-			num_q operator*(const int64_t &) const;
+			num_q operator*(const long long &) const;
 			num_q operator/(const num_q &) const;
 			num_q operator/(const num_z &) const;
-			num_q operator/(const int64_t &) const;
+			num_q operator/(const long long &) const;
 			num_q inverse() const { return num_q(this->_denominator, this->_numerator); };
 			num_q pow(const num_z &) const;
 
@@ -88,22 +88,22 @@
 
 			bool operator==(const num_q &) const;
 			bool operator==(const num_z &) const;
-			bool operator==(const int64_t &) const;
+			bool operator==(const long long &) const;
 			bool operator>=(const num_q &) const;
 			bool operator>=(const num_z &) const;
-			bool operator>=(const int64_t &) const;
+			bool operator>=(const long long &) const;
 			bool operator<=(const num_q &) const;
 			bool operator<=(const num_z &) const;
-			bool operator<=(const int64_t &) const;
+			bool operator<=(const long long &) const;
 			bool operator!=(const num_q &) const;
 			bool operator!=(const num_z &) const;
-			bool operator!=(const int64_t &) const;
+			bool operator!=(const long long &) const;
 			bool operator>(const num_q &) const;
 			bool operator>(const num_z &) const;
-			bool operator>(const int64_t &) const;
+			bool operator>(const long long &) const;
 			bool operator<(const num_q &) const;
 			bool operator<(const num_z &) const;
-			bool operator<(const int64_t &) const;
+			bool operator<(const long long &) const;
 	
 	};
