+++ src/num_q/attrib.cpp
@@ -15,7 +15,7 @@ num_q & num_q::operator=(const num_z &a){
 	return *this;
 }
 
-num_q & num_q::operator=(const int64_t &a){
+num_q & num_q::operator=(const long long &a){
 	this->_numerator = a;
 	this->_denominator = 1;
 	this->_sign = a < 0;
