+++ src/num_q/bool.cpp
@@ -8,7 +8,7 @@ bool num_q::operator==(const num_z &a) const {
 	return this->_denominator == 1 && this->_sign == a._sign && this->_numerator.abs_eq(a);
 }
 
-bool num_q::operator==(const int64_t &a) const {
+bool num_q::operator==(const long long &a) const {
 	return this->_denominator == 1 && this->_sign == (a < 0) && this->_numerator == (1 - 2*(a < 0)) * a ;
 }
 
@@ -31,7 +31,7 @@ bool num_q::operator>=(const num_z &a) const {
 	return res._numerator >= res._denominator;
 }
 
-bool num_q::operator>=(const int64_t &a) const {
+bool num_q::operator>=(const long long &a) const {
 	num_q res(*this);
 	res._numerator._sign = res._sign;
 	res._denominator *= a;
@@ -58,7 +58,7 @@ bool num_q::operator<=(const num_z &a) const {
 	return res._numerator <= res._denominator;
 }
 
-bool num_q::operator<=(const int64_t &a) const {
+bool num_q::operator<=(const long long &a) const {
 	num_q res(*this);
 	res._numerator._sign = res._sign;
 	res._denominator *= a;
@@ -74,7 +74,7 @@ bool num_q::operator!=(const num_z &a) const {
 	return !(*this == a);
 }
 
-bool num_q::operator!=(const int64_t &a) const {
+bool num_q::operator!=(const long long &a) const {
 	return !(*this == a);
 }
 
@@ -86,7 +86,7 @@ bool num_q::operator>(const num_z &a) const {
 	return !(*this <= a);
 }
 
-bool num_q::operator>(const int64_t &a) const {
+bool num_q::operator>(const long long &a) const {
 	return !(*this <= a);
 }
 
@@ -98,7 +98,7 @@ bool num_q::operator<(const num_z &a) const {
 	return !(*this >= a);
 }
 
-bool num_q::operator<(const int64_t &a) const {
+bool num_q::operator<(const long long &a) const {
 	return !(*this >= a);
 }
