+++ src/num_q/op_atrmul.cpp
@@ -19,7 +19,7 @@ num_q & num_q::operator*=(const num_z &a){
 	return *this;
 }
 
-num_q & num_q::operator*=(const int64_t &a){
+num_q & num_q::operator*=(const long long &a){
 	this->_sign ^= a < 0;
 	this->_numerator *= a;
