+++ src/num_z/op_atrmul.cpp
@@ -10,8 +10,8 @@ num_z & num_z::operator*=(const num_z &a){
 */	
 	int sign = this->_sign ^ a._sign;
 	uint32_t i, j, m, n;
-	uint64_t k;
-	uint64_t t;
+	unsigned long long k;
+	unsigned long long t;
 	num_z res(uint32_t(0));
 	
 	m = a._blocks;
@@ -24,7 +24,7 @@ num_z & num_z::operator*=(const num_z &a){
 	for(j = 0; j < m; ++j){
 		k = 0;
 		for(i = 0; i < n; ++i){
-			t = (uint64_t)this->_num[i] * (uint64_t)a._num[j] + res._num[i+j] + k;
+			t = (unsigned long long)this->_num[i] * (unsigned long long)a._num[j] + res._num[i+j] + k;
 			res._num[i + j] = t % _BASE_;
 			k = t / _BASE_;
 		}
@@ -53,11 +53,11 @@ num_z & num_z::operator*=(const mod_tuple &a){
 	return *this *= a.r;
 }
 
-num_z & num_z::operator*=(const int64_t &a){
+num_z & num_z::operator*=(const long long &a){
 	return *this *= (num_z)a;
 }
 
-num_z & num_z::operator*=(const uint64_t &a){
+num_z & num_z::operator*=(const unsigned long long &a){
 	return *this *= (num_z)a;
 }
