+++ src/num_z/op_atrsub.cpp
@@ -66,11 +66,11 @@ num_z & num_z::operator-=(const mod_tuple &a){
 	return *this -= a.r;
 }
 
-num_z & num_z::operator-=(const int64_t &a){
+num_z & num_z::operator-=(const long long &a){
 	return *this -= (num_z)a;
 }
 
-num_z & num_z::operator-=(const uint64_t &a){
+num_z & num_z::operator-=(const unsigned long long &a){
 	return *this -= (num_z)a;
 }
