+++ src/num_z/opsub.cpp
@@ -72,12 +72,12 @@ num_z num_z::operator-(const uint32_t &a) const {
 	return *this-res;
 }
 
-num_z num_z::operator-(const int64_t &a) const {
+num_z num_z::operator-(const long long &a) const {
 	num_z res(a);
 	return *this-res;
 }
 
-num_z num_z::operator-(const uint64_t &a) const {
+num_z num_z::operator-(const unsigned long long &a) const {
 	num_z res(a);
 	return *this-res;
 }
